requirements.txt
appdirs==1.4.3
chardet==3.0.3
click==6.7
cssselect==1.0.1
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
readability-lxml==0.6.2
requests==2.14.2
six==1.10.0
Werkzeug==0.12.2
